refactor: Move FakeSqsClient to fixtures
